chore: added security module
